🔨new (de)serialisation system fully integrated It will be much easier to implement more storage methods now
