+++ core/src/main/java/io/github/scroojalix/npcmanager/utils/storage/Storage.java
@@ -14,6 +14,7 @@ import io.github.scroojalix.npcmanager.utils.chat.Messages;
 import io.github.scroojalix.npcmanager.utils.npc.NPCData;
 import io.github.scroojalix.npcmanager.utils.storage.implementation.interfaces.RemoteStorage;
 import io.github.scroojalix.npcmanager.utils.storage.implementation.interfaces.StorageImplementation;
+import io.github.scroojalix.npcmanager.utils.storage.misc.JsonParser;
 
 public class Storage {
 
@@ -80,7 +81,7 @@ public class Storage {
                                 jsonFile.getParentFile().mkdirs();
                                 if (jsonFile.createNewFile()) {
                                     FileWriter writer = new FileWriter(jsonFile);
-                                    writer.write(data.toJson(true));
+                                    writer.write(JsonParser.toJson(data, true));
                                     writer.close();
                                 }
                             } catch (IOException e) {
@@ -154,7 +155,7 @@ public class Storage {
 				if (current.isFile() && current.getName().endsWith(".json")) {
 					try {
 						String json = new String(Files.readAllBytes(Paths.get(current.getPath())));
-						NPCData data = NPCData.fromJson(current.getName().replace(".json", ""), json, true);
+						NPCData data = JsonParser.fromJson(current.getName().replace(".json", ""), json, true);
 						if (data != null) {
 							boolean restore = true;
 							if (((RemoteStorage)implementation).isConnected()) {
