--- core/src/main/java/io/github/scroojalix/npcmanager/utils/storage/misc/ConfigurationSerializableAdapter.java
@@ -1,55 +1,0 @@
-package io.github.scroojalix.npcmanager.utils.storage.misc;
-
-import java.lang.reflect.Type;
-import java.util.LinkedHashMap;
-import java.util.Map;
-
-import com.google.gson.JsonDeserializationContext;
-import com.google.gson.JsonDeserializer;
-import com.google.gson.JsonElement;
-import com.google.gson.JsonParseException;
-import com.google.gson.JsonSerializationContext;
-import com.google.gson.JsonSerializer;
-import com.google.gson.reflect.TypeToken;
-
-import org.bukkit.configuration.serialization.ConfigurationSerializable;
-import org.bukkit.configuration.serialization.ConfigurationSerialization;
-
-public class ConfigurationSerializableAdapter implements JsonSerializer<ConfigurationSerializable>, JsonDeserializer<ConfigurationSerializable> {
-
-    final Type objectStringMapType = new TypeToken<Map<String, Object>>() {}.getType();
-
-    @Override
-    public ConfigurationSerializable deserialize(
-            final JsonElement json,
-            final Type typeOfT,
-            final JsonDeserializationContext context) throws JsonParseException
-    {
-        final Map<String, Object> map = new LinkedHashMap<>();
-
-        for (Map.Entry<String, JsonElement> entry : json.getAsJsonObject().entrySet()) {
-            final JsonElement value = entry.getValue();
-            final String name = entry.getKey();
-
-            if (value.isJsonObject() && value.getAsJsonObject().has(ConfigurationSerialization.SERIALIZED_TYPE_KEY)) {
-                map.put(name, this.deserialize(value, value.getClass(), context));
-            } else {
-                map.put(name, context.deserialize(value, Object.class));
-            }
-        }
-
-        return ConfigurationSerialization.deserializeObject(map);
-    }
-
-    @Override
-    public JsonElement serialize(
-            ConfigurationSerializable src,
-            Type typeOfSrc,
-            JsonSerializationContext context)
-    {
-        final Map<String, Object> map = new LinkedHashMap<>();
-        map.put(ConfigurationSerialization.SERIALIZED_TYPE_KEY, ConfigurationSerialization.getAlias(src.getClass()));
-        map.putAll(src.serialize());
-        return context.serialize(map, objectStringMapType);
-    }
-}
